requirements.txt
fastapi
uvicorn
pytest
SQLAlchemy>=2.0
psycopg2-binary
python-multipart
python-dotenv
httpx